requirements.txt
requests==2.31.0
pytest==7.4.4
allure-pytest==2.13.2